srv/agent/requirements.txt
# Busibox Agent API Requirements
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Database
psycopg2-binary==2.9.9

# Vector Database
pymilvus==2.3.5

# Object Storage
minio==7.2.3

# Queue
redis==5.0.1

# LLM Integration
litellm==1.17.0
openai==1.7.2

# Logging
structlog==24.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

